requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.9.0
fsspec==2023.4.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.2.1
networkx==3.0rc1
numpy==1.24.1
Pillow==10.1.0
requests==2.28.1
sympy==1.11.1
torch==2.2.0.dev20231127+cu121
torchaudio==2.2.0.dev20231127+cu121
torchvision==0.17.0.dev20231127+cu121
typing_extensions==4.8.0
urllib3==1.26.13
